Refactor ConfirmPasswordWalletRawTx
