+++ apps/wallet-mobile/src/features/Swap/useCases/StartSwapScreen/ListOrders/OpenOrders.tsx
@@ -29,10 +29,10 @@ import {useSelectedWallet} from '../../../../../SelectedWallet'
 import {COLORS} from '../../../../../theme'
 import {createRawTxSigningKey, generateCIP30UtxoCbor} from '../../../../../yoroi-wallets/cardano/utils'
 import {useTokenInfos, useTransactionInfos} from '../../../../../yoroi-wallets/hooks'
+import {ConfirmPasswordWalletRawTx} from '../../../common/ConfirmPasswordWalletRawTx/ConfirmPasswordWalletRawTx'
 import {Counter} from '../../../common/Counter/Counter'
 import {useNavigateTo} from '../../../common/navigation'
 import {PoolIcon} from '../../../common/PoolIcon/PoolIcon'
-import {ConfirmPasswordWalletRawTx} from '../../../common/RawTxConfirm/RawTxConfirm'
 import {useStrings} from '../../../common/strings'
 import {convertBech32ToHex, getMuesliSwapTransactionAndSigners, useCancellationOrderFee} from './helpers'
 import {mapOrders, MappedOrder} from './mapOrders'
